fix: move marketplace home to root
