+++ inlang/source-code/website/src/pages/Layout.tsx
@@ -109,9 +109,9 @@ function Header(props: { landingpage?: boolean; darkmode?: boolean; transparent?
 			},
 			{
 				name: `Missing something?`,
-				href; "https://github.com/orgs/inlang/discussions",
+				href: "https://github.com/orgs/inlang/discussions",
 				type: props.darkmode ? "textBackground" : ("text" as buttonType),
-			}
+			},
 		]
 	}
